Revert "Refactoring response to use HttpStatusCode instead of an int"
This reverts commit 9169c4e66745e546d768a26de356ca807a9cb13a.

+++ PactNet/Mocks/MockHttpService/Mappers/PactProviderServiceResponseMapper.cs
@@ -33,7 +33,7 @@ namespace PactNet.Mocks.MockHttpService.Mappers
 
             var to = new PactProviderServiceResponse
             {
-                Status = from.StatusCode,
+                Status = (int) from.StatusCode,
                 Headers = ConvertHeaders(from.Headers, from.Content)
             };
